+++ src/Compiler/Node.cpp
@@ -1,4 +1,4 @@
-#include <Ark/Parser/Node.hpp>
+#include <Ark/Compiler/Node.hpp>
 
 #include <termcolor.hpp>
